Перенесет рендеринг точек маршрута из main.js в controllers/trip-controller.js
